Created Itinerary model & Removed Example model
Added a new model and defined the Itinerary schema to include title, description, dateBegin, and dateEnd. Only the title is required. This is missing a reference to the day, which will need to be added after that model is defined.

Updated index.js to import the Itinerary schema and export it for use.