Authenication implemented
Used jwt token to created a token when a user is created and can be used to log a user in. Updated typeDefs and resolvers to handle adding and logging in user using authentication.

Created 2 simple pages Signup & Login to have a simple form to test the mutations. Added these destinations to the navigation and conditionally rendered logout if the user was already signed vs Signup/Login if the user was loggeed out.

Added an auth file in client folder to create the AuthService class for handling tokens