Fixed the addActivity mutation
The addActivity mutation was throwing an error.  In order to resolve it, I had to match the letter casing of the returned activity constant with the constant name.  In addition, I was having an issue querying activities because I was missing a closing curly bracket.  Also, my queries of all days and activities did not require parameters being passed in, so I removed those